Create knockout viewmodel for places and data-binds in view
JQuery listeners have been removed in favor of creating a knockout.js
View Model. Generally, methods like addPlace and removePlace will use a JQuery
ajax request to send the appropriate place info to the server, and the view
model will update in the callback once the ajax requests are complete.

To simplify views, the placeSearch.ejs template is deleted.