Create new authenticated homepage
When users are logged in, they will be redirected to a /my page that will
display their list of places and other user-related info. The intent is for a
cleaner separation between the landing page and the auth home page, to prevent
having to create complicated if(user)-type statements in the ejs view templates.