Update server api to return json instead of html
In the previous backend incarnation, the app made use of res.render() to
provide html partials with the correct list of places, created by using the
looping ability in the ejs templates. Now, the client handles displaying the
data correctly by using knockout, while the server is only responsible for
retrieving the place data in JSON format and returning it to the client.